API changes Party

+++ backend/server/inventory/party/party.controller.ts
@@ -1,16 +1,24 @@
 import * as bodyParser from 'body-parser';
 import {Router as expressRouter} from 'express';
-import { authorize } from '../../passport-util';
-import Party from './party.model';
+// Import { authorize } from '../../passport-util';
+import {PartyModel} from './party.model';
 import {Constants, PartyS} from 'fivebyone';
 import { PartyUtil } from './party.util';
+import { AuthUtil } from '../../util/auth.util';
 
-const {HTTP_OK, HTTP_BAD_REQUEST} = Constants;
+const {HTTP_OK, HTTP_BAD_REQUEST, HTTP_UNAUTHORIZED} = Constants;
 
 const router = expressRouter();
 
 const listParty = async(_request: any, response: any) => {
 
+  const sessionDetails = AuthUtil.findSessionDetails(_request);
+  if (!sessionDetails.company) {
+
+    return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+  }
+  const Party = PartyModel.createModel(sessionDetails.company);
   const taxes = await Party.find();
   return response.status(HTTP_OK).json(taxes);
 
@@ -21,6 +29,13 @@ const getParty = async(request: any, response: any) => {
 
   try {
 
+    const sessionDetails = AuthUtil.findSessionDetails(request);
+    if (!sessionDetails.company) {
+
+      return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+    }
+    const Party = PartyModel.createModel(sessionDetails.company);
     const party = await Party.findById(request.params.id);
     if (!party) {
 
@@ -41,6 +56,14 @@ const saveParty = async(request: any, response: any) => {
 
   try {
 
+    const sessionDetails = AuthUtil.findSessionDetails(request);
+    if (!sessionDetails.company) {
+
+      return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+    }
+    const Party = PartyModel.createModel(sessionDetails.company);
+
     const party = new Party(request.body);
     await party.save();
     return response.status(HTTP_OK).json(party);
@@ -58,6 +81,14 @@ const updateParty = async(request: any, response: any) => {
   try {
 
     const {id} = request.params;
+    const sessionDetails = AuthUtil.findSessionDetails(request);
+    if (!sessionDetails.company) {
+
+      return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+    }
+    const Party = PartyModel.createModel(sessionDetails.company);
+
     const party = await Party.findById(id);
 
     const updateObject: PartyS = request.body;
@@ -80,6 +111,13 @@ const deleteParty = async(request: any, response: any) => {
   try {
 
     const {id} = request.params;
+    const sessionDetails = AuthUtil.findSessionDetails(request);
+    if (!sessionDetails.company) {
+
+      return response.status(HTTP_UNAUTHORIZED).json('Permission denied.');
+
+    }
+    const Party = PartyModel.createModel(sessionDetails.company);
     const resp = await Party.deleteOne({_id: id});
     if (resp.deletedCount === 0) {
 
@@ -97,10 +135,10 @@ const deleteParty = async(request: any, response: any) => {
 
 };
 
-router.route('/').get(authorize, listParty);
-router.route('/:id').get(authorize, getParty);
-router.route('/').post(authorize, bodyParser.json(), saveParty);
-router.route('/:id').put(authorize, bodyParser.json(), updateParty);
-router.route('/:id')['delete'](authorize, deleteParty);
+router.route('/').get(AuthUtil.authorize, listParty);
+router.route('/:id').get(AuthUtil.authorize, getParty);
+router.route('/').post(AuthUtil.authorize, bodyParser.json(), saveParty);
+router.route('/:id').put(AuthUtil.authorize, bodyParser.json(), updateParty);
+router.route('/:id')['delete'](AuthUtil.authorize, deleteParty);
 
 export default router;
